pyproject.toml
[project]
name = "monobankdaily"
version = "0.1.0"
description = "Telegram bot for daily Monobank spending reports"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[ext]>=22.1",
    "sqlalchemy>=2.0",
    "python-dotenv>=1.0.0",
    "pytz>=2024.1",
    "httpx>=0.27.0",
    "cryptography>=44.0.0",
    "alembic>=1.14.0",
]

[project.scripts]
monobankdaily = "src.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "ty>=0.0.1a8",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
    "B024",   # abstract base class without abstract methods (design pattern)
    "SIM105", # contextlib.suppress (readability preference)
    "SIM108", # ternary operator (readability preference)
    "E721",   # type comparison (needed for media type check)
    "ARG001", # unused function args (pytest fixtures need this)
    "ARG002", # unused method args (pytest fixtures need this)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
